post-rootfs: Make async-commit and /info/ optional
For minimize rootfs size.

+++ common/scripts/post-async-commit.sh
@@ -4,6 +4,8 @@ POST_ROOTFS_ONLY=1
 
 source "${RK_POST_HELPER:-$(dirname "$(realpath "$0")")/../post-hooks/post-helper}"
 
+[ "$RK_ROOTFS_ASYNC_COMMIT" ] || return 0
+
 message "Installing async-commit service..."
 
 rm -f etc/init.d/S*_commit.sh \
